Consistent values cleanup types precision

+++ example/Dambreak2dMDBC.jl
@@ -2,7 +2,7 @@ using SPHExample
 
 let
     Dimensions = 2
-    FloatType  = Float32
+    FloatType  = Float64
 
     SimConstantsDambreak = SimulationConstants{FloatType}(dx=0.01,c₀=88.14487860902641, δᵩ = 0.1, CFL=0.5, α = 0.01)
